64bit work merging from V5 Macintosh Cocoa Squeak VM microsecond & utc clock udates
DropPlugin.h, sqFilePlubinBasic.h,hostWindowPlugin.h,SecurityPlugin.h,
SoundPlugin  64bit work

SocketPlugin.h
add IPV6 declares

sqFFI.h
missing pushBoolean define

sq.h
Add ioMicroSeconds,ioUtcWithOffset
Add EventTypeComplex
readImageFromFileHeapSizeStartingAt desriredHeapSize 32bit clean
id ->indexNumber to make Objective-C routines happy (id is type not identifier)

sqNamedPrims.c
64bit work

sqMacUnixInterfaceSound.c
sqUnixSoundMacOSX.c
64bit work, but this is removed in V5 64bit work anyway.

sqMacUUID.c
64bit work, but this is replaced by unix code in V5 64bit work anyway

sqMacMain.c
Mac Carbon 4.2.4b1

sqMacNSPluginUILogic2.c
move sq.h to avoid issues with undefined usqInt

sqMacTime.c
add utc primitive

sqMacWindowUniversal.c
ioShowDisplayOnWindow char *

sqPlatformSpecific.h
Add microsecond and utc stuff


git-svn-id: http://squeakvm.org/svn/squeak/trunk@2158 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c11c041b73f1fb67ebd896b580a6477a25cc22d6

+++ FilePlugin/sqFilePluginBasicPrims.c
@@ -166,7 +166,7 @@ sqInt sqFileInit(void) {
 	   Should be called once at startup time.
 	*/
 #if VM_PROXY_MINOR > 6
-	thisSession = interpreterProxy->getThisSessionID();
+	thisSession = (int) interpreterProxy->getThisSessionID();
 #else
 	thisSession = ioLowResMSecs() + time(NULL);
 	if (thisSession == 0) thisSession = 1;	/* don't use 0 */
